bugfix
Former-commit-id: 27deb6501cac9f678b422dcfcd2adf3740b3bd86 [formerly 8f4b20f75774dd0dd4f83f8e765a63a666e9bbfe]
Former-commit-id: ac73dca1b8fca4eafae171506c9f635d45742936

+++ src/examples/step-lubrication-sine/LubricationSine.cpp
@@ -31,7 +31,7 @@ shared_ptr<Triangulation<2> > LubricationSine::makeGrid(double L,
 	shared_ptr<Triangulation<2> > rect = make_shared<Triangulation<2> >();
 	dealii::Point<2> x1(0,1);
 	dealii::Point<2> x2(0,1);
-	std::vector<size_t> repetitions;
+	std::vector<unsigned int> repetitions;
 	repetitions.at(0) = L / averageHeight;
 	repetitions.at(1) = 1;
 	bool colorize = true; 	// set boundary ids automatically to
